+++ src/core/gt_logmodel.cpp
@@ -19,6 +19,7 @@
 #include "gt_logmodel.h"
 #include "gt_logging.h"
 #include "gt_logdest.h"
+#include "gt_utilities.h"
 
 void
 GtLogSignalSlotDestination::write(std::string const& message,
@@ -36,28 +37,6 @@ GtLogSignalSlotDestination::write(std::string const& message,
             QDateTime::fromSecsSinceEpoch(t)
         }
     );
-}
-
-GtLogModel&
-GtLogModel::instance()
-{
-    static GtLogModel self;
-    return self;
-}
-
-QString
-GtLogModel::format(QString const& msg, int level, GtLogDetails const& details)
-{
-    Q_UNUSED(level);
-
-    auto time = details.time.toString("hh:mm:ss");
-
-    if (!details.id.isEmpty())
-    {
-        return QStringLiteral("[%1] [%2] %3").arg(time, details.id, msg);
-    }
-
-    return QStringLiteral("[%1] %2").arg(time, msg);
 }
 
 GtLogModel::GtLogModel() :
@@ -82,29 +61,191 @@ GtLogModel::GtLogModel() :
     }
 }
 
+GtLogModel&
+GtLogModel::instance()
+{
+    static GtLogModel self;
+    return self;
+}
+
+QString
+GtLogModel::format(QString const& msg, GtLogDetails const& details)
+{
+    auto time = details.time.toString("hh:mm:ss");
+
+    if (!details.id.isEmpty())
+    {
+        return QStringLiteral("[%1] [%2] %3").arg(time, details.id, msg);
+    }
+
+    return QStringLiteral("[%1] %2").arg(time, msg);
+}
+
+QString
+GtLogModel::format(const Entry& entry)
+{
+    return format(entry.msg, entry.details);
+}
+
+int
+GtLogModel::rowCount(const QModelIndex& parent) const
+{
+    return m_entries.count();
+}
+
+int
+GtLogModel::columnCount(const QModelIndex& parent) const
+{
+    // level, time, id, message
+    return 4;
+}
+
+QVariant
+GtLogModel::data(const QModelIndex& index, int role) const
+{
+    int row = index.row();
+    int col = index.column();
+    if (row < 0 || col < 0 || row >= m_entries.count())
+    {
+        return {};
+    }
+
+    // delegate DisplayRole role to custom roles
+    if (role == Qt::DisplayRole)
+    {
+        role = columnToRole(col);
+    }
+
+    Entry const& entry = m_entries[index.row()];
+
+    // get data
+    switch (role)
+    {
+    case LevelRole:
+        return entry.level;
+    case TimeRole:
+        return entry.details.time.toString("hh:mm:ss");
+    case IdRole:
+        return entry.details.id;
+    case MessageRole:
+        return entry.msg;
+    default:
+        break;
+    }
+
+    return {};
+}
+
+QVariant
+GtLogModel::headerData(int section, Qt::Orientation orientation, int role) const
+{
+    if (orientation != Qt::Horizontal)
+    {
+        return {};
+    }
+
+    // delegate DisplayRole role to custom roles
+    if (role == Qt::DisplayRole)
+    {
+        role = columnToRole(section);
+    }
+
+    switch (role)
+    {
+    case LevelRole:
+        return tr("Level");
+    case TimeRole:
+        return tr("Time");
+    case IdRole:
+        return tr("ID");
+    case MessageRole:
+        return tr("Message");
+    default:
+        return {};
+    }
+}
+
+bool
+GtLogModel::exportLogToFile(const QString& fileName)
+{
+    QFile file(fileName);
+
+    if (!file.open(QFile::WriteOnly))
+    {
+        gtError() << tr("Could not export output log! File: '%1'")
+                     .arg(fileName);
+        return false;
+    }
+
+    QTextStream out(&file);
+
+    for (Entry const& entry : qAsConst(m_entries))
+    {
+        auto level = gt::log::levelFromInt(entry.level);
+
+        out << gt::log::levelToString(level).c_str() << ' '
+            << format(entry) << "\r\n";
+    }
+
+    return true;
+}
+
 QMimeData*
 GtLogModel::mimeData(QModelIndexList const& indexes) const
 {
-    QString str;
+    QStringList list;
+    int lastRow = -1;
 
+    /// we assume a sorted list
     for (const QModelIndex& index : indexes)
     {
         int row = index.row();
         if (row < 0 || row >= m_entries.count())
         {
-            continue;
+            return {};
         }
 
-        Entry const& entry = m_entries.at(row);
-        const auto level = gt::log::levelFromInt(entry.level);
+        // append new row
+        if (lastRow != row)
+        {
+            list.append(QString{});
+            lastRow = row;
+        }
 
-        str += QStringLiteral("%1 %2\r\n")
-                   .arg(gt::log::levelToString(level).c_str(), format(entry));
+        // access last string
+        assert(!list.empty());
+        QString& str = list.last();
+
+        auto const& entry = m_entries.at(row);
+        auto const  level = gt::log::levelFromInt(entry.level);
+
+        // append data
+        switch (columnToRole(index.column()))
+        {
+        case LevelRole:
+            str.append(QString::fromStdString(gt::log::levelToString(level)));
+            break;
+        case TimeRole:
+            str.append(entry.details.time.toString("[hh:mm:ss]"));
+            break;
+        case IdRole:
+            if (!entry.details.id.isEmpty())
+            {
+                str.append(gt::quoted(entry.details.id, QChar{'['}, QChar{']'}));
+            }
+            break;
+        case MessageRole:
+            str.append(entry.msg);
+            break;
+        default:
+            continue;
+        }
+        // padding
+        str.append(QChar{' '});
     }
 
-    QMimeData* mimeData = new QMimeData;
-    mimeData->setText(str);
-
+    auto* mimeData = new QMimeData;
+    mimeData->setText(list.join(QStringLiteral("\r\n")));
     return mimeData;
 }
 
@@ -119,64 +260,10 @@ GtLogModel::setMaxLogLength(int val)
     m_maxEntries = val;
 }
 
-int
-GtLogModel::rowCount(const QModelIndex& parent) const
-{
-    return parent.isValid() ? 0 : m_entries.count();
-}
-
-QVariant
-GtLogModel::data(const QModelIndex& index, int role) const
-{
-    if (index.row() < 0 || index.row() >= m_entries.count())
-    {
-        return {};
-    }
-
-    Entry const& entry = m_entries[index.row()];
-
-    switch (role)
-    {
-    case Qt::DisplayRole:
-    {
-        return format(entry);
-    }
-    case GtLogModel::LogLevelRole:
-        return entry.level;
-    }
-
-    return {};
-}
-
-bool
-GtLogModel::exportLogToFile(const QString& filename)
-{
-    QFile file(filename);
-
-    if (!file.open(QFile::WriteOnly))
-    {
-        gtError() << "could not open file! (" << filename << ")";
-        return false;
-    }
-
-    QTextStream out(&file);
-
-    for (Entry const& entry : qAsConst(m_entries))
-    {
-        auto level = gt::log::levelFromInt(entry.level);
-
-        out << gt::log::levelToString(level).c_str() << "\t"
-            << entry.msg << "\r\n";
-    }
-
-    return true;
-}
-
 void
 GtLogModel::onMessage(const QString& msg, int level, Details const& details)
 {
-    m_mutex.lock();
-
+    QMutexLocker locker{&m_mutex};
     if (m_timer.isActive())
     {
         m_tmpEntries << Entry({ msg, level, details});
@@ -186,8 +273,6 @@ GtLogModel::onMessage(const QString& msg, int level, Details const& details)
         insertMessage(msg, level, details);
         m_timer.start(500);
     }
-
-    m_mutex.unlock();
 }
 
 void
@@ -199,6 +284,7 @@ GtLogModel::execClear()
         m_entries.clear();
         endResetModel();
         m_tmpClearLog = false;
+        emit logCleared();
     }
 }
 
@@ -257,6 +343,18 @@ GtLogModel::removeElementList(QModelIndexList indexList, int first, int last)
     endRemoveRows();
 }
 
+QModelIndex
+GtLogModel::index(int row, int column, const QModelIndex& parent) const
+{
+    return createIndex(row, column);
+}
+
+QModelIndex
+GtLogModel::parent(const QModelIndex& index) const
+{
+    return {};
+}
+
 void
 GtLogModel::insertQueue()
 {
@@ -295,7 +393,7 @@ GtLogModel::insertQueue()
 
     if (delta > 0)
     {
-        beginRemoveRows(QModelIndex(), 0, delta - 1);
+        beginRemoveRows(QModelIndex{}, 0, delta - 1);
         for (int i = 0; i < delta; i++)
         {
             m_entries.removeFirst();
@@ -303,8 +401,7 @@ GtLogModel::insertQueue()
         endRemoveRows();
     }
 
-    beginInsertRows(QModelIndex(), m_entries.size(), m_entries.size() + newRows);
-
+    beginInsertRows(QModelIndex{}, m_entries.size(), m_entries.size() + newRows);
     for (Entry const& entry : qAsConst(m_tmpEntries))
     {
         m_entries << entry;